--- tests/test_middleware.py
@@ -1,43 +1,0 @@
-from unittest.mock import patch
-import pytest
-
-from labs.api.types import GithubModel
-from labs.litellm_service.request import RequestLiteLLM
-from labs.middleware import call_llm_with_context
-
-
-class TestCallLlmWithContext:
-    # Successfully calls the LLM with the provided context and returns the expected output
-    @patch("labs.middleware.vectorize_and_find_similar")
-    @patch.object(RequestLiteLLM, "completion_without_proxy")
-    @patch("labs.middleware.call_agent_to_apply_code_changes")
-    def test_successful_llm_call_with_context(
-        self, mocked_agent, mocked_completion, mocked_embeddings
-    ):
-        # Mocking dependencies
-        mocked_embeddings.return_value = [["file1", "/path/to/file1", "content1"]]
-        mocked_completion.return_value = (
-            "model",
-            {"choices": [{"message": {"content": "response"}}]},
-        )
-        mocked_agent.return_value = ["file1"]
-
-        github = GithubModel(github_token="token", repo_owner="owner", repo_name="repo")
-        issue_summary = "Fix the bug in the authentication module"
-
-        success, result = call_llm_with_context(github, issue_summary)
-
-        assert success
-        assert result == ("model", {"choices": [{"message": {"content": "response"}}]})
-
-    # Ensure the GithubModel is instantiated correctly with the correct parameters
-
-    def test_empty_summary(self):
-        github = GithubModel(github_token="token", repo_owner="owner", repo_name="repo")
-        issue_summary = ""
-
-        with pytest.raises(Exception) as excinfo:
-            call_llm_with_context(github, issue_summary)
-
-        assert "issue_summary cannot be empty" in str(excinfo.value)
-        # Corrects the mocking of find_similar_embeddings attribute in the rag module
